Fixed script issues with selling
BaseInventory has the needed net vars, Test tube redone, no longer uses a dictionary, got rid of improper function calling. Fixed SampleSO. No longer needed a min max, got rid of the unnecessary * 100 and cast to int which broke the samples range. Replaced SampleSOs string with an enum to work with network. Inventory looks at the correct stuff. Did an override for testtube to allow it to pass through the proper info for SampleData and so that is "CanBeSold" is dynamic and not based off the ItemSO