Gameplay event system created. Added to event bus
Sadly the gameplay events will be used for audio as there was not plan earlier for this system. This will help a lot in the future.
Added two new functions to the event bus. PublishGameplayEvent which works only for events of type GameplayEvent. This is so we can subscribe to all events of this type for mass logic.
And PublishForType. This is like PublishGameplayEvent but much more generic. Untested at the moment but likely will work
Created a structure for gameplay events. Added a middleman to handle logic from events and handle relaying. Other stuff probably